Handle failures in sysctl BPF helpers
Mostly related truncation when the buffer for either the sysctl name or
sysctl value is not large enough. Add a new field to record that
truncation has occurred and add a note when logging the sysctl change.

errno.h has to included after vmlinux.h, so reorder the includes of
sysctl-logger.bpf.c a bit.