Enable BPF verifier logging in verbose mode
BPF verifier logging is on when the BPF program fails to load, but it
can be equally useful to see how a BPF program passes the verifier for
compare and contrast with the failed one.

To do this we have to switch to sysctl_logger_bpf__open_opts() instead
of using the sysctl_logger_bpf__open_and_load() call, allocate
bpf_object_open_opts struct, and set it's size.