Remove newline at the end of sysctl value
With this the logging is prettier and does not spanning across multiple
lines.

This is done inside the BPF program since the return of
bpf_sysctl_get_{current,new}_value() is the exact length of the string
value. The slight caveat here is that we need to switch the type of ret
from int to long to avoid verifier being confused by the combination of
register splitting and long to int conversion through shifting.