Move newline trimming of sysctl value to userspace
Instead of dealing with possible verifier error, opt for the less
performant/efficient string processing in userspace where we copy the
whole sysctl_logger_event struct then modify the string.

Sysctl is generally not something that is performance critical, so
moving the string processing into userspace should be much of an issue.