Notify systemd when setup is finished
This should help ensure that systemd wait for sysctl-logger to finish
it's setup (i.e. loading and attaching the BPF program) before
proceeding to start other services so we don't loss any sysctl changes.

Alternatively we could opt for Type=forking instead and not take
dependency on libsystemd itself. But that requires more lines of changes
so I'm choosed the (seemly) easier approach for now.