Change service type to exec from notify
A colleague pointed out that for a simple monitoring program, using
Type=notify may be a bit of an overkill, requiring taking libsystemd as
a dependency. Thus revert commit e27113b28c63 "Notify systemd when setup
is finished" for now, and change service type to exec, which should
already buy enough time for the BPF program to be loaded.