Rephrase logging to clarify what event represents
Currently sysctl-logger captures the initiation of a sysctl value
change, but does not check whether the change actually went through, so
it is misleading to log each event as if they succeed.

While is should be possible to detect whether sysctl change succeed with
an additional kretprobe, it is not yet done, so instead change to
logging to make it not as misleading for now.